+++ OPHD/Technology/TechnologyCatalog.cpp
@@ -1,4 +1,4 @@
-#include "TechnologyReader.h"
+#include "TechnologyCatalog.h"
 #include "../XmlSerializer.h"
 
 #include <NAS2D/Utility.h>
