+++ OPHD/ophd.vcxproj.filters
@@ -354,7 +354,7 @@
     <ClCompile Include="Population\PopulationTable.cpp">
       <Filter>Source Files\Population</Filter>
     </ClCompile>
-    <ClCompile Include="Technology\TechnologyReader.cpp">
+    <ClCompile Include="Technology\TechnologyCatalog.cpp">
       <Filter>Source Files\Technology</Filter>
     </ClCompile>
   </ItemGroup>
@@ -773,7 +773,7 @@
     <ClInclude Include="Technology\Technology.h">
       <Filter>Header Files\Technology</Filter>
     </ClInclude>
-    <ClInclude Include="Technology\TechnologyReader.h">
+    <ClInclude Include="Technology\TechnologyCatalog.h">
       <Filter>Header Files\Technology</Filter>
     </ClInclude>
   </ItemGroup>
